+++ subgraph/schema.graphql
@@ -1,7 +1,7 @@
 type Lock @entity {
   id: ID!
   address: Bytes!
-  lockManagers: [address!]
+  lockManagers: [Bytes!]
   keys: [Key!] @derivedFrom(field: "lock")
   createdAtBlock: BigInt
   version: BigInt
